backend/qaemdecor/pyproject.toml
[tool.poetry]
name = "qaemdecor"
version = "0.1.0"
description = "this is qaem decor backend web api"
authors = ["Mahmood D Monkabadi <m.mahmooddm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2.5"
psycopg2-binary = "^2.9.1"
Pillow = "^8.3.0"
django-debug-toolbar = "^3.2.1"
djangorestframework = "^3.12.4"
django-rest-auth = "^0.9.5"
django-filter = "^2.4.0"
django-oauth-toolkit = "^1.5.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
